Little update to installation file
This changes are needed for many Windows installations that have permissions restrictions on the home folder that only allow to read/write/create on containing files/folders and do not allow changes to the folder itself.

In these cases the "is_writable" PHP function fail and the WP installation become ugly for some users.
So, better to check if the sample config file is writable and if we can create the config file instead of simply testing the writability of the folder.

P.S. Windows understand both but some plugins or 3th party application doesn't. So, even if only for code clarity, PHP have a directory separator constant... use it!